backend/db.py
import os

from dotenv import load_dotenv
from supabase import Client, create_client


BASE_DIR = os.path.dirname(__file__)
load_dotenv(os.path.join(BASE_DIR, ".env"))

SUPABASE_URL = os.getenv("SUPABASE_URL")
SUPABASE_KEY = os.getenv("SUPABASE_KEY")

if not SUPABASE_URL or not SUPABASE_KEY:
    raise RuntimeError("SUPABASE_URL and SUPABASE_KEY must be set in backend/.env or environment variables.")

supabase: Client = create_client(SUPABASE_URL, SUPABASE_KEY)
