backend/requirements.txt
fastapi>=0.100.0
uvicorn
supabase
python-dotenv
pydantic>=2.0.0
